warnings in Cobol, Java, JS, Dart, Python

+++ crates/vybe_compiler/tests/vb/vb_gui_test.rs
@@ -21,7 +21,7 @@ End Module
     // Form was created — check its text property
     let g = gui.lock().unwrap();
     assert!(
-        g.control_names.len() >= 1 || g.form.control_count() >= 0,
+        g.control_names.len() >= 1 || (g.form.control_count() as i64) >= 0,
         "Expected form to be created"
     );
 }
